TST Fixes test

+++ sklearn/preprocessing/tests/test_encoders.py
@@ -321,7 +321,7 @@ def test_X_is_not_1D_pandas(method):
      [['A', 'B'], ['cat']], np.object_),
     (np.array([['A', 'cat'], ['B', 'cat']]),
      [['A', 'B'], ['cat']], np.str_),
-    (np.array([[1, 2], [np.nan, 2]]), [[1, np.nan], [2]], np.integer),
+    (np.array([[1, 2], [np.nan, 2]]), [[1, np.nan], [2]], np.float_),
     (np.array([['A', np.nan], [None, np.nan]], dtype=object),
      [['A', None], [np.nan]], np.object_),
     ], ids=['mixed', 'numeric', 'object', 'string', 'missing-float',
